+++ changelog.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docstrings for methods list_users, list_stocks, get_stock, add_stock_price,
 - Lots of placeholder methods
 - create_game method now returns status
+- Discord bot framework
+
 
 ### Fixed
 - Data format for get_stock method
@@ -25,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - datetime_updated can no longer be null in stock_picks table (SQLite DB)
 - username to display_name in add_user method
 - get_stock and list_stock_prices methods now use _sql_get()
+- Renamed StockGame class to Backend
 
 ### Removed
 - Misc placeholders and completed todos